+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.17.0] - 2024-02-06
 ### Added
 * Added `output_format` parameter for document upload function, that indicates
   the file extension of the desired file format for the translated document. 
@@ -284,7 +284,7 @@ Version increased to avoid conflicts with old packages on PyPI.
 Initial version.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.16.1...HEAD
+[1.17.0]: https://github.com/DeepLcom/deepl-python/compare/v1.16.1...v1.17.0
 [1.16.1]: https://github.com/DeepLcom/deepl-python/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/DeepLcom/deepl-python/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/DeepLcom/deepl-python/compare/v1.14.0...v1.15.0